Damaged State Implementation
this is by far the most unkempt push for this project but I will say that it is very duplicatable. Enemies now respond to being attacked with a stunned duration and a movement away from the enemy in the direction of the attack. I split the implementation up between the health and enemy movement classes. This implementation is called in the Attack area a class where the stun duration is decided.